// packages/core/src/domain/types.ts
export type BodyPartId =
  | 'head' | 'neck'
  | 'torso_front' | 'torso_back' | 'torso_side_l' | 'torso_side_r'
  | 'arm_upper_l' | 'arm_lower_l' | 'hand_l'
  | 'arm_upper_r' | 'arm_lower_r' | 'hand_r'
  | 'leg_upper_l' | 'leg_lower_l' | 'foot_l'
  | 'leg_upper_r' | 'leg_lower_r' | 'foot_r'
  | 'groin';